+++ ct/ct.vcxproj
@@ -135,6 +135,7 @@
     <ClCompile Include="CodeIncrement.cpp" />
     <ClCompile Include="CodeSet.cpp" />
     <ClCompile Include="CodeTypes.cpp" />
+    <ClCompile Include="CodeWarning.cpp" />
     <ClCompile Include="CtModule.cpp" />
     <ClCompile Include="Cxx.cpp" />
     <ClCompile Include="CxxArea.cpp" />
@@ -148,6 +149,7 @@
     <ClCompile Include="CxxString.cpp" />
     <ClCompile Include="CxxSymbols.cpp" />
     <ClCompile Include="CxxToken.cpp" />
+    <ClCompile Include="Editor.cpp" />
     <ClCompile Include="Interpreter.cpp" />
     <ClCompile Include="Lexer.cpp" />
     <ClCompile Include="Library.cpp" />
@@ -167,6 +169,7 @@
     <ClInclude Include="CodeIncrement.h" />
     <ClInclude Include="CodeSet.h" />
     <ClInclude Include="CodeTypes.h" />
+    <ClInclude Include="CodeWarning.h" />
     <ClInclude Include="CtModule.h" />
     <ClInclude Include="Cxx.h" />
     <ClInclude Include="CxxArea.h" />
@@ -181,6 +184,7 @@
     <ClInclude Include="CxxString.h" />
     <ClInclude Include="CxxSymbols.h" />
     <ClInclude Include="CxxToken.h" />
+    <ClInclude Include="Editor.h" />
     <ClInclude Include="Interpreter.h" />
     <ClInclude Include="Lexer.h" />
     <ClInclude Include="Library.h" />
